Implemented basic chapter 2 image and refactored doubles to floats
Following the 2018 NVIDIA technical blog tutorial: https://developer.nvidia.com/blog/accelerated-ray-tracing-cuda/

Worked through creating that first image as a sanity check, and updated all doubles to floats since doubles do not perform well on GPU

+++ src/CUDAInOneWeekend/color.h
@@ -17,7 +17,7 @@
 using color = vec3;
 
 
-inline double linear_to_gamma(double linear_component)
+inline float linear_to_gamma(float linear_component)
 {
     if (linear_component > 0)
         return std::sqrt(linear_component);
